+++ src/app/locales/en/components/qos-params-editor.js
@@ -1,0 +1,4 @@
+export default {
+  header: 'QOS parameters',
+  newValueKey: 'New key-value...',
+};