+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-5493 Added QOS parameters to storages
 * VFS-3335 Added storage edition and deletion
 * VFS-4640 Added auto-polling of space occupancy, disabled join space in emergency, changed session expired message
 * VFS-4596 Major changes to support unified and standalone Onepanel GUI